convert context to singleton
