+++ packages/gatsby/cache-dir/gatsby-browser-entry.js
@@ -1,93 +1,15 @@
-import React from "react"
-import PropTypes from "prop-types"
 import Link, {
   withPrefix,
   withAssetPrefix,
   navigate,
-  push,
-  replace,
-  navigateTo,
   parsePath,
 } from "gatsby-link"
 import { useScrollRestoration } from "gatsby-react-router-scroll"
+import { StaticQuery, useStaticQuery } from "./static-query"
 import PageRenderer from "./public-page-renderer"
 import loader from "./loader"
 
 const prefetchPathname = loader.enqueue
-
-const StaticQueryContext = React.createContext({})
-
-function StaticQueryDataRenderer({ staticQueryData, data, query, render }) {
-  const finalData = data
-    ? data.data
-    : staticQueryData[query] && staticQueryData[query].data
-
-  return (
-    <React.Fragment>
-      {finalData && render(finalData)}
-      {!finalData && <div>Loading (StaticQuery)</div>}
-    </React.Fragment>
-  )
-}
-
-const StaticQuery = props => {
-  const { data, query, render, children } = props
-
-  return (
-    <StaticQueryContext.Consumer>
-      {staticQueryData => (
-        <StaticQueryDataRenderer
-          data={data}
-          query={query}
-          render={render || children}
-          staticQueryData={staticQueryData}
-        />
-      )}
-    </StaticQueryContext.Consumer>
-  )
-}
-
-const useStaticQuery = query => {
-  if (
-    typeof React.useContext !== `function` &&
-    process.env.NODE_ENV === `development`
-  ) {
-    throw new Error(
-      `You're likely using a version of React that doesn't support Hooks\n` +
-        `Please update React and ReactDOM to 16.8.0 or later to use the useStaticQuery hook.`
-    )
-  }
-  const context = React.useContext(StaticQueryContext)
-
-  // query is a stringified number like `3303882` when wrapped with graphql, If a user forgets
-  // to wrap the query in a grqphql, then casting it to a Number results in `NaN` allowing us to
-  // catch the misuse of the API and give proper direction
-  if (isNaN(Number(query))) {
-    throw new Error(`useStaticQuery was called with a string but expects to be called using \`graphql\`. Try this:
-
-import { useStaticQuery, graphql } from 'gatsby';
-
-useStaticQuery(graphql\`${query}\`);
-`)
-  }
-
-  if (context[query]?.data) {
-    return context[query].data
-  } else {
-    throw new Error(
-      `The result of this StaticQuery could not be fetched.\n\n` +
-        `This is likely a bug in Gatsby and if refreshing the page does not fix it, ` +
-        `please open an issue in https://github.com/gatsbyjs/gatsby/issues`
-    )
-  }
-}
-
-StaticQuery.propTypes = {
-  data: PropTypes.object,
-  query: PropTypes.string.isRequired,
-  render: PropTypes.func,
-  children: PropTypes.func,
-}
 
 function graphql() {
   throw new Error(
@@ -106,7 +28,6 @@ export {
   parsePath,
   navigate,
   useScrollRestoration,
-  StaticQueryContext,
   StaticQuery,
   PageRenderer,
   useStaticQuery,
